Refactor to isolate common txn storage encodings
