+++ server/src/engine/core/model/mod.rs
@@ -279,7 +279,7 @@ impl Model {
             }
             // since we've locked this down, no one else can parallely create another model in the same space (or remove)
             if G::FS_IS_NON_NULL {
-                let mut txn_driver = global.namespace_txn_driver().lock();
+                let mut txn_driver = global.gns_driver().lock();
                 // prepare txn
                 let txn = gns::model::CreateModelTxn::new(
                     SpaceIDRef::new(&space_name, &space),
@@ -294,7 +294,7 @@ impl Model {
                     model.get_uuid(),
                 )?;
                 // commit txn
-                match txn_driver.try_commit(txn) {
+                match txn_driver.gns_driver().try_commit(txn) {
                     Ok(()) => {}
                     Err(e) => {
                         // failed to commit, request cleanup
@@ -358,7 +358,7 @@ impl Model {
                     model.delta_state().schema_current_version().value_u64(),
                 ));
                 // commit txn
-                global.namespace_txn_driver().lock().try_commit(txn)?;
+                global.gns_driver().lock().gns_driver().try_commit(txn)?;
                 // request cleanup
                 global.purge_model_driver(
                     space_name,
