+++ server/src/engine/storage/common_encoding/r1/impls/gns/mod.rs
@@ -25,104 +25,25 @@
 */
 
 use {
-    super::super::raw::journal::{JournalAdapter, JournalWriter},
     crate::{
         engine::{
             core::GlobalNS,
             data::uuid::Uuid,
-            error::{RuntimeResult, TransactionError},
+            error::RuntimeResult,
             mem::BufferedScanner,
-            storage::{
-                common_encoding::r1::{self, PersistObject},
-                safe_interfaces::{FSInterface, LocalFS},
-            },
-            txn::{gns, SpaceIDRef},
+            storage::common_encoding::r1::{self, PersistObject},
+            txn::SpaceIDRef,
         },
         util::EndianQW,
     },
     std::marker::PhantomData,
 };
 
-mod model;
-mod space;
+pub mod model;
+pub mod space;
 // test
 #[cfg(test)]
 mod tests;
-
-/// The GNS transaction driver is used to handle DDL transactions
-pub struct GNSTransactionDriverAnyFS<Fs: FSInterface = LocalFS> {
-    journal: JournalWriter<Fs, GNSAdapter>,
-}
-
-impl<Fs: FSInterface> GNSTransactionDriverAnyFS<Fs> {
-    pub fn new(journal: JournalWriter<Fs, GNSAdapter>) -> Self {
-        Self { journal }
-    }
-    pub fn into_inner(self) -> JournalWriter<Fs, GNSAdapter> {
-        self.journal
-    }
-    pub fn __journal_mut(&mut self) -> &mut JournalWriter<Fs, GNSAdapter> {
-        &mut self.journal
-    }
-    /// Attempts to commit the given event into the journal, handling any possible recovery triggers and returning
-    /// errors (if any)
-    pub fn try_commit<GE: GNSEvent>(&mut self, gns_event: GE) -> RuntimeResult<()> {
-        let mut buf = vec![];
-        buf.extend(GE::OPC.to_le_bytes());
-        GE::encode_super_event(gns_event, &mut buf);
-        self.journal
-            .append_event_with_recovery_plugin(GNSSuperEvent(buf.into_boxed_slice()))?;
-        Ok(())
-    }
-}
-
-/*
-    journal implementor
-*/
-
-/// the journal adapter for DDL queries on the GNS
-#[derive(Debug)]
-pub struct GNSAdapter;
-
-impl JournalAdapter for GNSAdapter {
-    const RECOVERY_PLUGIN: bool = true;
-    type JournalEvent = GNSSuperEvent;
-    type GlobalState = GlobalNS;
-    type Error = crate::engine::fractal::error::Error;
-    fn encode(GNSSuperEvent(b): Self::JournalEvent) -> Box<[u8]> {
-        b
-    }
-    fn decode_and_update_state(payload: &[u8], gs: &Self::GlobalState) -> RuntimeResult<()> {
-        if payload.len() < 2 {
-            return Err(TransactionError::DecodedUnexpectedEof.into());
-        }
-        macro_rules! dispatch {
-            ($($item:ty),* $(,)?) => {
-                [$(<$item as GNSEvent>::decode_and_update_global_state),*, |_, _| Err(TransactionError::DecodeUnknownTxnOp.into())]
-            };
-        }
-        static DISPATCH: [fn(&mut BufferedScanner, &GlobalNS) -> RuntimeResult<()>; 9] = dispatch!(
-            gns::space::CreateSpaceTxn,
-            gns::space::AlterSpaceTxn,
-            gns::space::DropSpaceTxn,
-            gns::model::CreateModelTxn,
-            gns::model::AlterModelAddTxn,
-            gns::model::AlterModelRemoveTxn,
-            gns::model::AlterModelUpdateTxn,
-            gns::model::DropModelTxn
-        );
-        let mut scanner = BufferedScanner::new(&payload);
-        let opc = unsafe {
-            // UNSAFE(@ohsayan):
-            u16::from_le_bytes(scanner.next_chunk())
-        };
-        match DISPATCH[(opc as usize).min(DISPATCH.len())](&mut scanner, gs) {
-            Ok(()) if scanner.eof() => return Ok(()),
-            Ok(_) => Err(TransactionError::DecodeCorruptedPayloadMoreBytes.into()),
-            Err(e) => Err(e),
-        }
-    }
-}
 
 /*
     Events
@@ -133,8 +54,6 @@ impl JournalAdapter for GNSAdapter {
     ---
     [OPC:2B][PAYLOAD]
 */
-
-pub struct GNSSuperEvent(Box<[u8]>);
 
 /// Definition for an event in the GNS (DDL queries)
 pub trait GNSEvent
