+++ server/src/engine/storage/common_encoding/r1/mod.rs
@@ -29,6 +29,7 @@
 //! This is revision 1 of high-level interface encoding.
 //!
 
+pub mod impls;
 pub mod map;
 pub mod obj;
 // tests
