+++ server/src/engine/storage/v1/raw/journal/raw.rs
@@ -45,7 +45,7 @@
 use crate::engine::storage::common::interface::fs_traits::FileOpen;
 
 use {
-    super::{rw::SDSSFileIO, spec::Header},
+    super::super::{rw::SDSSFileIO, spec::Header},
     crate::{
         engine::{
             error::{RuntimeResult, StorageError},
